Fix: MenuList를 UI 폴더로 옮기고 반응형 상태바 기능 수정
